Refactor: Moved schemas to seperate folder
