+++ src/app/login/forgot-password/page.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { z } from 'zod'
+import { zodResolver } from '@hookform/resolvers/zod'
+import { useForm } from 'react-hook-form'
+import { toast } from 'sonner'
+import { useRouter } from 'next/navigation'
+
+export const page = () => {
+  return <div>page</div>
+}
+
+export default page
