+++ src/app/login/login.action.ts
@@ -1,12 +1,12 @@
 'use server'
 
 import { z } from 'zod'
-import { signUpSchema } from './SignUpForm'
+import { signInSchema } from '@/schemas/zod/schemas'
 import { prisma } from '@/lib/prisma'
 import { Argon2id } from 'oslo/password'
 import { lucia } from '@/lib/lucia'
 import { cookies } from 'next/headers'
-import { signInSchema } from './LoginForm'
+import { signUpSchema } from '@/schemas/zod/schemas'
 import { redirect } from 'next/navigation'
 import { generateState } from 'arctic'
 import { generateCodeVerifier } from 'oslo/oauth2'
